Add files via upload
This is the program that allows to make multi-mailing, e.g. invitation for the wedding or any other event. 
You will have the opportunity to put the list of the guests you want to invite in the separate file (._invited_names.txt). The message itself will be typed in the file ._starting_letter.txt.
When running the main.py the program will create separate invitations for each guest, including his name.
enjoy!