Add files via upload
This is a very interesting quiz allowing to study all 50 states of the USA. 
You are to write the name of the state. If the state exists and was not mentioned before - the player get 1 point of 50. If you have no ideas any more - the game ends and create the list of the states that were not mentioned. So you will have a chance to learn them!
enjoy