Add files via upload
This is a great program to generate a complicated passwords and store them locally on your computer in .txt file. 
The passwords are systematized and are indicated in correspondence with site name (or other location) they are relevant to and username/email. 
When generated the password gets copied to your LOCAL clipboard and is ready to pe pasted to your site

Enjoy!

tkinter, pyperclip, messagebox