Create Stock Trading and Telegram News Alert Project
The user inputs the Stock name and the actual Company name. As well as his API keys for www.alphavantage.co, newsapi.org and Telegram chat ID. 
The program shows the closing price of the bonds for yesterday and its delta compared to the previous day. It also picks up the most popular news about the company from yesterday. 
Bot sends all this information to the defined Telegram chat.